ai-video-bot/requirements.txt
google-generativeai>=0.7.0
moviepy>=1.0.3
Pillow>=10.0.0
requests>=2.31.0
gTTS>=2.3.2
instagrapi>=2.0.0
google-auth-oauthlib>=1.0.0
google-api-python-client>=2.100.0
numpy>=1.24.0

